requirements.txt
anyio==4.8.0
asgiref==3.8.1
attrs==25.1.0
certifi==2025.1.31
chardet==5.2.0
charset-normalizer==3.4.1
colorama==0.4.6
colorlog==6.9.0
Django==4.2
django-cors-headers==4.7.0
django-environ==0.12.0
django-extensions==3.2.3
django-filter==24.3
djangorestframework==3.14.0
djangorestframework_simplejwt==5.4.0
drf-spectacular==0.28.0
drf-yasg==1.21.9
exceptiongroup==1.2.2
geographiclib==2.0
geopy==2.4.1
gunicorn==23.0.0
h11==0.14.0
httpcore==1.0.7
httpx==0.28.1
icalendar==6.1.1
idna==3.10
inflection==0.5.1
jsonschema==4.23.0
jsonschema-specifications==2024.10.1
logging==0.4.9.6
packaging==24.2
pillow==11.1.0
psycopg==3.2.4
psycopg-binary==3.2.4
psycopg2==2.9.10
PyJWT==2.10.1
python-dateutil==2.9.0.post0
python-dotenv==1.0.1
python-telegram-bot==21.4
pytz==2025.1
PyYAML==6.0.2
qrcode==8.0
referencing==0.36.2
reportlab==4.3.1
requests==2.32.3
rpds-py==0.22.3
six==1.17.0
sniffio==1.3.1
sqlparse==0.5.3
typing_extensions==4.12.2
tzdata==2025.1
uritemplate==4.1.1
urllib3==2.3.0
whitenoise==6.9.0
